Add theme.
